lab2/requirements.txt
torch==2.10.0
torchvision==0.25.0
numpy==2.2.6
pandas==2.3.3
Pillow==12.1.1
matplotlib==3.10.8
tqdm==4.67.3